# docs/swagger.yaml
definitions:
...
  api.MLAPIConfig:
    properties:
      algorithms:
        items:
          type: string
        type: array
      batch_size:
        type: integer
      drift_detection:
        type: boolean
      enable_continuous:
        type: boolean
      enabled:
        type: boolean
      min_confidence:
        type: number
      retrain_threshold:
        type: integer
      training_interval:
        type: string
      validation_ratio:
        type: number
      voting_strategy:
        type: string
      weights:
        additionalProperties:
          format: float64
          type: number
        type: object
    type: object
...
  api.MLStatusResponse:
    properties:
      buffer_size:
        type: integer
      config:
        $ref: '#/definitions/api.MLAPIConfig'
      enabled:
        type: boolean
      is_running:
        type: boolean
      last_training:
        type: string
      performance_history:
        items:
          $ref: '#/definitions/ml.TrainingPerformance'
        type: array
      sample_count:
        type: integer
    type: object
...
  ml.TrainingPerformance:
    properties:
      algorithmMetrics:
        additionalProperties:
          format: float64
          type: number
        type: object
      driftDetected:
        type: boolean
      sampleCount:
        format: int64
        type: integer
      timestamp:
        type: string
      trainingDuration:
        $ref: '#/definitions/time.Duration'
      validationScore:
        format: float64
        type: number
    type: object
...
  time.Duration:
    enum:
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    format: int64
    type: integer
    x-enum-varnames:
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
info:
  contact: {}
paths:
...
  /api/v1/ml/config:
    get:
      description: Returns current ML system configuration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.MLAPIConfig'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get ML configuration
      tags:
      - ml
    put:
      consumes:
      - application/json
      description: Update ML system configuration
      parameters:
      - description: ML configuration
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/api.MLAPIConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update ML configuration
      tags:
      - ml
...
  /api/v1/ml/performance:
    get:
      description: Returns historical performance metrics for the ML anomaly detection
        system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ml.TrainingPerformance'
            type: array
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get ML performance history
      tags:
      - ml
  /api/v1/ml/status:
    get:
      description: Returns current status and statistics of the ML anomaly detection
        system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.MLStatusResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get ML system status
      tags:
      - ml
  /api/v1/ml/train:
    post:
      description: Manually trigger ML model training with current data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Force ML model training
      tags:
      - ml
...
swagger: "2.0"